pyproject.toml
[tool.poetry]
name = "goodfire"
version = "0.3.5"
description = "Understand and steer your AI models"
authors = ["Goodfire"]
license = "Apache-2.0"
readme = "README.md"
classifiers = [
  "Typing :: Typed",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
  "Operating System :: POSIX",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Microsoft :: Windows",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License"
]

[tool.poetry.dependencies]
python = "^3.10"
ipywidgets = "^8.1.5"
httpx = "^0.27.2"
pydantic = "^2.9.2"
scipy = "^1.14.1"
scikit-learn = "^1.6.0"
numpy = ">= 1.26.4"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
